1. Move files of shuffler. 2. Compatibility of startup arguments.
