+++ jvm/core/main/scala/com/webank/eggroll/core/clustermanager/session/SessionManager.scala
@@ -32,13 +32,13 @@ import scala.collection.mutable
 import scala.collection.mutable.ArrayBuffer
 
 object SessionManager {
-  private val activeSessions = new ConcurrentHashMap[String, ErServerSessionDeployment]()
+  private val activeSessions = new ConcurrentHashMap[String, ErSessionDeployment]()
 
   // (sessionId, (bindingId, ErProcessorBatch))
   private val activeSessionBound = new ConcurrentHashMap[String, ConcurrentHashMap[String, ErProcessorBatch]]()
 
   activeSessions.put(StringConstants.UNKNOWN,
-    ErServerSessionDeployment(
+    ErSessionDeployment(
       id = StringConstants.UNKNOWN,
       serverCluster = ErServerCluster(),
       rolls = Array.empty,
@@ -61,7 +61,7 @@ object SessionManager {
         val rolls = deployer.createRolls()
         val eggs = deployer.createEggs()
 
-        val newDeployment = ErServerSessionDeployment(
+        val newDeployment = ErSessionDeployment(
           id = sessionId,
           serverCluster = healthyCluster,
           rolls = rolls.processors,
@@ -114,7 +114,7 @@ object SessionManager {
     })
 
     // todo: find and populate serverCluster
-    val newDeployment = ErServerSessionDeployment(
+    val newDeployment = ErSessionDeployment(
       id = sessionId,
       serverCluster = serverCluster,
       rolls = rolls.toArray,
@@ -132,7 +132,7 @@ object SessionManager {
     else null
   }
 
-  def addSession(sessionId: String, deployment: ErServerSessionDeployment): Unit = {
+  def addSession(sessionId: String, deployment: ErSessionDeployment): Unit = {
     if (activeSessions.contains(sessionId)) {
       throw new IllegalArgumentException(s"sessionId ${sessionId} already exists")
     }
@@ -149,7 +149,7 @@ object SessionManager {
     null
   }
 
-  def getSessionDeployment(sessionId: String): ErServerSessionDeployment = {
+  def getSessionDeployment(sessionId: String): ErSessionDeployment = {
     activeSessions.get(sessionId)
   }
 }
