+++ jvm/core/main/scala/com/webank/eggroll/core/schedule/TaskPlan.scala
@@ -47,4 +47,4 @@ class SampleTaskPlan(uri: CommandURI, job: ErJob) extends BaseTaskPlan(uri, job)
 class FilterTaskPlan(uri: CommandURI, job: ErJob) extends BaseTaskPlan(uri, job)
 class SubtractByKeyTaskPlan(uri: CommandURI, job: ErJob) extends BaseTaskPlan(uri, job)
 class UnionTaskPlan(uri: CommandURI, job: ErJob) extends BaseTaskPlan(uri, job)
-class PutBatchTaskPlan(uri: CommandURI, job: ErJob) extends BaseTaskPlan(uri, job)
+class PutAllTaskPlan(uri: CommandURI, job: ErJob) extends BaseTaskPlan(uri, job)
